draft some utility functions
