+++ pyretroicor/util.py
@@ -1,0 +1,1 @@
+"""I/O and file path senity check"""
